Refactor draft_pick to move function to lib
* Refactor draft_pick to move function to lib
* Will enable functions to be used in validations
* Closes #531
