+++ lib/ex338_web/templates/draft_pick/index.html.eex
@@ -4,7 +4,7 @@
 
 <section class="draft-picks-container">
   <div class="draft-picks-collection">
-    <%= render "team_summary_table.html", conn: @conn, current_user: @current_user, draft_picks: @draft_picks %>
+    <%= render "team_summary_table.html", conn: @conn, current_user: @current_user, fantasy_teams: @fantasy_teams %>
     <%= render "table.html", conn: @conn, current_user: @current_user, draft_picks: @draft_picks %>
   </div>
 </section>
